ajustado as visualizações dos usuários nos respectivos tempos de eleição

+++ app/Http/Controllers/Admin/eleicaoController.php
@@ -28,7 +28,7 @@ class eleicaoController extends Controller
         }
 
         return view('admin.eleicao.index', [
-            'eleicoes' => $eleicoes->paginate(3),
+            'eleicoes' => $eleicoes->paginate(5),
             'search' => isset($request->search) ? $request->search : ''
         ]);
     }
@@ -40,7 +40,7 @@ class eleicaoController extends Controller
 
     public function store(eleicaoRequest $request)
     {
-        $data = $request->validated();        
+        $data = $request->validated();
 
         // JUNTANDO DATA E HORA DA ELEIÇÃO E INSCRIÇÃO
         $data['start_date_eleicao'] .= ' '.$data['start_time_eleicao'];
@@ -52,17 +52,17 @@ class eleicaoController extends Controller
         $validator = Validator::make(request()->all(), []);
         if($data['end_date_inscricao'] <= $data['start_date_inscricao']){
             if($data['end_date_eleicao'] <= $data['start_date_eleicao']){
-                $validator->errors()->add('end_time_inscricao', 'Horario final deve ser maior ou igual ao inicial')->add('end_time_eleicao', 'Horario final deve ser maior ou igual ao inicial');
+                $validator->errors()->add('end_time_inscricao', 'Horario final deve ser maior que o inicial')->add('end_time_eleicao', 'Horario final deve ser maior que o inicial');
                 return back()->withErrors($validator)->withInput();
             }else{
-                $validator->errors()->add('end_time_inscricao', 'Horario final deve ser maior ou igual ao inicial');
+                $validator->errors()->add('end_time_inscricao', 'Horario final deve ser maior que o inicial');
                 return back()->withErrors($validator)->withInput();
             }
         } else if($data['end_date_eleicao'] <= $data['start_date_eleicao']){
-            $validator->errors()->add('end_time_eleicao', 'Horario final deve ser maior ou igual ao inicial');
+            $validator->errors()->add('end_time_eleicao', 'Horario final deve ser maior que o inicial');
             return back()->withErrors($validator)->withInput();
         }
-        
+
 
         // REMOVENDO HORA ELEIÇÃO E INSCRIÇÃO DA VARIAVEL
         unset($data['start_time_eleicao']);
@@ -92,10 +92,14 @@ class eleicaoController extends Controller
 
         return view('admin.eleicao.show', [
             'eleicoes' => $eleicao,
-            'eleicaoStartDateHasPassed' => EleicaoService::eleicaoStartDateHasPassed($eleicao),
-            'eleicaoEndDateHasPassed' => EleicaoService::eleicaoEndDateHasPassed($eleicao),
-            'inscricaoStartDateHasPassed' => EleicaoService::inscricaoStartDateHasPassed($eleicao),
-            'inscricaoEndDateHasPassed' => EleicaoService::inscricaoEndDateHasPassed($eleicao),
+
+            'beforeInscricao' => EleicaoService::beforeInscricao($eleicao),
+            'duringInscricao' => EleicaoService::duringInscricao($eleicao),
+            'afterInscricao' => EleicaoService::afterInscricao($eleicao),
+            'beforeEleicao' => EleicaoService::beforeEleicao($eleicao),
+            'duringEleicao' => EleicaoService::duringEleicao($eleicao),
+            'afterEleicao' => EleicaoService::afterEleicao($eleicao),
+
             'total' => $total,
             'vencedor' => $vencedor
         ]);
@@ -140,14 +144,14 @@ class eleicaoController extends Controller
         $validator = Validator::make(request()->all(), []);
         if($data['end_date_inscricao'] <= $data['start_date_inscricao']){
             if($data['end_date_eleicao'] <= $data['start_date_eleicao']){
-                $validator->errors()->add('end_time_inscricao', 'Horario final deve ser maior ou igual ao inicial')->add('end_time_eleicao', 'Horario final deve ser maior ou igual ao inicial');
+                $validator->errors()->add('end_time_inscricao', 'Horario final deve ser maior que o inicial')->add('end_time_eleicao', 'Horario final deve ser maior que o inicial');
                 return back()->withErrors($validator)->withInput();
             }else{
-                $validator->errors()->add('end_time_inscricao', 'Horario final deve ser maior ou igual ao inicial');
+                $validator->errors()->add('end_time_inscricao', 'Horario final deve ser maior que o inicial');
                 return back()->withErrors($validator)->withInput();
             }
         } else if($data['end_date_eleicao'] <= $data['start_date_eleicao']){
-            $validator->errors()->add('end_time_eleicao', 'Horario final deve ser maior ou igual ao inicial');
+            $validator->errors()->add('end_time_eleicao', 'Horario final deve ser maior que o inicial');
             return back()->withErrors($validator)->withInput();
         }
 
@@ -164,6 +168,10 @@ class eleicaoController extends Controller
 
     public function destroy(Eleicao $eleicao)
     {
+        if(EleicaoService::duringEleicao($eleicao)){
+            return back()->with('warning', 'Eleição em andamento, não pode excluir');
+        }
+
         $eleicao->delete();
 
         return redirect()->route('admin.eleicao.index')->with('success', 'Eleição removida com sucesso');
